@@ -1,19 +1,23 @@
-package sirtapp.sirtdetection.com.sirtapp;
+package sirtapp.sirtdetection.com.sirtapp.utils.images;
 
 import android.graphics.Bitmap;
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.widget.ImageView;
 
+import androidx.appcompat.app.AppCompatActivity;
+import sirtapp.sirtdetection.com.sirtapp.R;
+
 public class Result extends AppCompatActivity {
-static ImageView result,image;
+    static ImageView result, image;
 
-public Result(ImageView image){
-    this.image=image;
-}
-public Result(){
+    public Result(ImageView image) {
+        this.image = image;
+    }
 
-}
+    public Result() {
+
+    }
+
     public static void mostar(Bitmap bitmap) {
         result.setImageBitmap(bitmap);
     }
@@ -23,11 +27,5 @@ public Result(){
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_result);
         result = findViewById(R.id.image2);
-    }
-
-    public void mostrar() {
-        image.buildDrawingCache();
-        Bitmap bmap = image.getDrawingCache();
-        result.setImageBitmap(bmap);
     }
 }
